ingestion/requirements.txt
pandas
kafka-python
pyarrow